# run build job on bigger flavor