--- a/.yamato/standalone-build-test.yml
+++ b/.yamato/standalone-build-test.yml
@@ -6,7 +6,7 @@ test_linux_standalone_{{ editor.version }}:
   agent:
     type: Unity::VM
     image: package-ci/ubuntu:stable
-    flavor: b1.medium
+    flavor: b1.large
   variables:
     UNITY_VERSION: {{ editor.version }}
   commands: