### Proposed change(s)
baseline (from https://github.com/Unity-Technologies/ml-agents/pull/4767)
[2018.4: 17m 04s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/switch-yamato-linux/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2018.4/4752496/job)
[2019.4: 14m 15s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/switch-yamato-linux/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2019.4/4752495/job)
[2020.1: 12m 23s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/switch-yamato-linux/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2020.1/4752494/job)
[2020.2: 12m 49s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/switch-yamato-linux/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2020.2/4752493/job)

b1.large
[2018.4: 11m 11s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/standalone-build-speedup/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2018.4/4753165/job)
[2019.4:  9m 02s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/standalone-build-speedup/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2019.4/4753164/job)
[2020.1:  9m 07s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/standalone-build-speedup/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2020.1/4753163/job)
[2020.2:  8m 00s](https://yamato.cds.internal.unity3d.com/jobs/497-ml-agents/tree/standalone-build-speedup/.yamato%252Fstandalone-build-test.yml%2523test_linux_standalone_2020.2/4753162/job)


### Useful links (Github issues, JIRA tickets, ML-Agents forum threads etc.)



### Types of change(s)

- [ ] Bug fix
- [ ] New feature
- [ ] Code refactor
- [ ] Breaking change
- [ ] Documentation update
- [ ] Other (please describe)

### Checklist
- [ ] Added tests that prove my fix is effective or that my feature works
- [ ] Updated the [changelog](https://github.com/Unity-Technologies/ml-agents/blob/master/com.unity.ml-agents/CHANGELOG.md) (if applicable)
- [ ] Updated the [documentation](https://github.com/Unity-Technologies/ml-agents/tree/master/docs) (if applicable)
- [ ] Updated the [migration guide](https://github.com/Unity-Technologies/ml-agents/blob/master/docs/Migrating.md) (if applicable)

### Other comments